backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openpyxl==3.1.2
pypdf2==3.0.1
python-dotenv==1.0.0
alembic==1.12.1
Pillow==10.3.0
imagehash==4.3.1
numpy==1.26.4
opencv-python==4.10.0.84
boto3==1.40.31
pytest>=7.0.0
requests>=2.28.0

